+++ src/test/java/de/michab/scream/fcos/SchemeCharacterTest.java
@@ -3,7 +3,7 @@
  *
  * Copyright © 1998-2022 Michael G. Binz
  */
-package de.michab.scream;
+package de.michab.scream.fcos;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
@@ -11,8 +11,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
 
-import de.michab.scream.fcos.FirstClassObject;
-import de.michab.scream.fcos.SchemeCharacter;
+import de.michab.scream.ScreamBaseTest;
 
 public class SchemeCharacterTest
 {
