+++ src/test/java/de/michab/scream/fcos/SchemeStringTest.java
@@ -3,7 +3,7 @@
  *
  * Copyright © 1998-2022 Michael G. Binz
  */
-package de.michab.scream;
+package de.michab.scream.fcos;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
@@ -14,11 +14,10 @@ import static org.junit.jupiter.api.Assertions.fail;
 
 import org.junit.jupiter.api.Test;
 
+import de.michab.scream.RuntimeX;
+import de.michab.scream.ScreamBaseTest;
+import de.michab.scream.ScreamException;
 import de.michab.scream.ScreamException.Code;
-import de.michab.scream.fcos.Cons;
-import de.michab.scream.fcos.SchemeCharacter;
-import de.michab.scream.fcos.SchemeDouble;
-import de.michab.scream.fcos.SchemeString;
 
 public class SchemeStringTest
 {
