+++ src/test/java/de/michab/scream/fcos/SymbolTest.java
@@ -3,14 +3,15 @@
  *
  * Copyright © 1998-2022 Michael G. Binz
  */
-package de.michab.scream;
+package de.michab.scream.fcos;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import org.junit.jupiter.api.Test;
 
-import de.michab.scream.fcos.Symbol;
+import de.michab.scream.ScreamBaseTest;
+import de.michab.scream.ScreamEvaluator;
 
 public class SymbolTest extends ScreamBaseTest
 {
