+++ src/test/java/de/michab/scream/fcos/VectorTest.java
@@ -3,18 +3,17 @@
  *
  * Copyright © 1998-2022 Michael G. Binz
  */
-package de.michab.scream;
+package de.michab.scream.fcos;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.fail;
 
 import org.junit.jupiter.api.Test;
 
+import de.michab.scream.RuntimeX;
+import de.michab.scream.ScreamBaseTest;
+import de.michab.scream.ScreamException;
 import de.michab.scream.ScreamException.Code;
-import de.michab.scream.fcos.Cons;
-import de.michab.scream.fcos.FirstClassObject;
-import de.michab.scream.fcos.SchemeBoolean;
-import de.michab.scream.fcos.Vector;
 
 public class VectorTest extends ScreamBaseTest
 {
